# api-docs.yaml
openapi: 3.0.0
paths:
...
  /api/v1/users/profile:
    get:
      operationId: UsersController_getProfile
      summary: Get current user profile
      parameters: []
      responses:
        '200':
          description: Profile retrieved successfully
      tags: &ref_3
        - users
      security: &ref_4
        - bearer: []
    put:
      operationId: UsersController_updateProfile
      summary: Update current user profile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileDto'
      responses:
        '200':
          description: Profile updated successfully
      tags: *ref_3
      security: *ref_4
  /api/v1/users/{id}/public:
    get:
      operationId: UsersController_getPublicProfile
      summary: Get public user profile
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Public profile retrieved successfully
        '404':
          description: User not found
      tags: *ref_3
      security: *ref_4
  /api/v1/users/test:
    get:
      operationId: UsersController_test
      summary: Test endpoint
      parameters: []
      responses:
        '200':
          description: Test successful
      tags: *ref_3
      security: *ref_4
  /api/v1/users/admin/{id}/block:
    put:
      operationId: UsersController_blockUser
      summary: Block a user (Admin only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: User blocked successfully
      tags: *ref_3
      security: *ref_4
  /api/v1/users/admin/{id}/unblock:
    put:
      operationId: UsersController_unblockUser
      summary: Unblock a user (Admin only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: User unblocked successfully
      tags: *ref_3
      security: *ref_4
...
info:
  title: memi.az API
  description: Recommerce Platform Backend API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UpdateProfileDto:
      type: object
      properties: {}
...